+++ packages/libs/datastore/src/main.ts
@@ -14,7 +14,6 @@ export * from "./source-map/source-map";
 export { Lazy, LazyPromise } from "./lazy";
 export * from "./parsing/text-utility";
 
-export * from "./parsing/yaml";
 export * from "./source-map/blaming";
 export * from "./parsing/stable-object";
 export * from "./processor";
