+++ packages/libs/yaml/src/utils.ts
@@ -1,59 +1,25 @@
 import { YamlNodeWithPath } from "./parser";
 import { Kind, YamlMap, YamlMapping, YamlNode, YamlSequence } from "./types";
 
-// /**
-//  * List all the child nodes of the providede root.
-//  * @param yamlAstNode Root.
-//  * @param currentPath
-//  */
-// export function* listYamlAstDecendants(yamlAstNode: YamlNode, currentPath: string[] = []): Iterable<YamlNodeWithPath> {
-//   const todos: YamlNodeWithPath[] = [{ path: currentPath, node: yamlAstNode }];
-//   let todo: YamlNodeWithPath | undefined;
-//   while ((todo = todos.pop())) {
-//     // report self
-//     yield todo;
-
-//     // traverse
-//     if (todo.node) {
-//       switch (todo.node.kind) {
-//         case Kind.MAPPING:
-//           {
-//             todos.push({ node: todo.node.value, path: todo.path.concat([todo.node.key.value]) });
-//           }
-//           break;
-//         case Kind.MAP:
-//           for (const mapping of todo.node.mappings) {
-//             todos.push({ node: mapping, path: todo.path });
-//           }
-//           break;
-//         case Kind.SEQ:
-//           {
-//             for (let i = 0; i < todo.node.items.length; ++i) {
-//               todos.push({ node: todo.node.items[i], path: todo.path.concat([i]) });
-//             }
-//           }
-//           break;
-//       }
-//     }
-//   }
-// }
-
-export function walkYamlAst(yamlAstNode: YamlNode, walker: (node: YamlNodeWithPath) => void) {
-  walkYamlAstInternal(yamlAstNode, [], walker);
+/**
+ * Walk the yaml ast tree.
+ * @param node Node to start walking.
+ * @param walker Callback called for each nodes
+ */
+export function walkYamlAst(node: YamlNode, walker: (node: YamlNodeWithPath) => void) {
+  walkYamlAstInternal(node, [], walker);
 }
 
-export function walkYamlAstInternal(
+function walkYamlAstInternal(
   node: YamlNode,
   currentPath: Array<string | number>,
   walker: (node: YamlNodeWithPath) => void,
-) {
+): void {
   walker({ path: currentPath, node });
 
   switch (node.kind) {
     case Kind.MAPPING:
-      {
-        walkYamlAstInternal(node.value, currentPath.concat([node.key.value]), walker);
-      }
+      return walkYamlAstInternal(node.value, currentPath.concat([node.key.value]), walker);
       break;
     case Kind.MAP:
       for (const mapping of node.mappings) {
@@ -61,10 +27,8 @@ export function walkYamlAstInternal(
       }
       break;
     case Kind.SEQ:
-      {
-        for (let i = 0; i < node.items.length; ++i) {
-          walkYamlAstInternal(node.items[i], currentPath.concat([i]), walker);
-        }
+      for (let i = 0; i < node.items.length; ++i) {
+        walkYamlAstInternal(node.items[i], currentPath.concat([i]), walker);
       }
       break;
   }
